requirements.txt
sqlalchemy==1.3.23
gino==1.0.1
aiogram==2.9
python-dotenv==0.15.0
asyncpg==0.22.0
